updated variable assignment output
I realised that I listed the "system internal" notation of negated variables  in the output variable assignment, which also resulted in entries like "{..., -7: False, ...}". (Which is unnecessary since the assignment to False makes the "-" notation superfluous) I changed it so that instead it returns the absolute value of the variable!