added visualisation
Using the parameter "-dpll_visual" a tree is now plotted that visualises how the search space is systematically tried with different trees. To make it possible to observe the changes, the algorithm is slowed down considerably and might therefore run for a very long time if the search space is big.